requirements.txt
discord.py==2.3.2
aiohttp==3.9.1
pyyaml==6.0.1
requests==2.31.0
